add: 【正常系】todoList, uniqueIdが更新されること、addInputValueがリセットされること
